fixes & cleanup
