+++ core/runtime.rs
@@ -17,6 +17,7 @@ use crate::modules::ModuleMap;
 use crate::op_void_async;
 use crate::op_void_sync;
 use crate::ops::*;
+use crate::snapshot_util::get_context_data;
 use crate::source_map::SourceMapCache;
 use crate::source_map::SourceMapGetter;
 use crate::Extension;
@@ -436,69 +437,6 @@ impl JsRuntime {
     let mut snapshot_extensions = vec![];
     let mut module_map_data = None;
     let mut module_handles = vec![];
-
-    fn get_context_data(
-      scope: &mut v8::HandleScope<()>,
-      context: v8::Local<v8::Context>,
-    ) -> (
-      Vec<String>,
-      Vec<v8::Global<v8::Module>>,
-      v8::Global<v8::Object>,
-    ) {
-      fn data_error_to_panic(err: v8::DataError) -> ! {
-        match err {
-          v8::DataError::BadType { actual, expected } => {
-            panic!(
-              "Invalid type for snapshot data: expected {expected}, got {actual}"
-            );
-          }
-          v8::DataError::NoData { expected } => {
-            panic!("No data for snapshot data: expected {expected}");
-          }
-        }
-      }
-
-      let mut module_handles = vec![];
-      let mut scope = v8::ContextScope::new(scope, context);
-
-      let extensions: Vec<String> =
-        match scope.get_context_data_from_snapshot_once::<v8::Value>(0) {
-          Ok(val) => serde_v8::from_v8(&mut scope, val).unwrap(),
-          Err(err) => data_error_to_panic(err),
-        };
-
-      // The 1st element is the module map itself, followed by X number of module
-      // handles. We need to deserialize the "next_module_id" field from the
-      // map to see how many module handles we expect.
-      match scope.get_context_data_from_snapshot_once::<v8::Object>(1) {
-        Ok(val) => {
-          let next_module_id = {
-            let info_str = v8::String::new(&mut scope, "info").unwrap();
-            let info_data: v8::Local<v8::Array> = val
-              .get(&mut scope, info_str.into())
-              .unwrap()
-              .try_into()
-              .unwrap();
-            info_data.length()
-          };
-
-          for i in 2..=next_module_id {
-            match scope
-              .get_context_data_from_snapshot_once::<v8::Module>(i as usize)
-            {
-              Ok(val) => {
-                let module_global = v8::Global::new(&mut scope, val);
-                module_handles.push(module_global);
-              }
-              Err(err) => data_error_to_panic(err),
-            }
-          }
-
-          (extensions, module_handles, v8::Global::new(&mut scope, val))
-        }
-        Err(err) => data_error_to_panic(err),
-      }
-    }
 
     let (mut isolate, snapshot_options) = if options.will_snapshot {
       let (snapshot_creator, snapshot_loaded) =
